+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -8,12 +8,16 @@ import javax.validation.constraints.Size;
 
 @Data
 public class Film {
-    private int id;
+    private long id;
+
     @NotBlank(message = "Name shouldn't be empty")
     private String name;
+
     @Size(max = 200, message = "Description should be less than 200 symbols")
     private String description;
+
     private String releaseDate;
+
     @Positive(message = "Duration should be positive")
     private int duration;
 }
