+++ src/main/java/ru/yandex/practicum/filmorate/model/User.java
@@ -7,11 +7,14 @@ import javax.validation.constraints.Pattern;
 
 @Data
 public class User {
-    private int id;
+    private long id;
+
     @Email
     private String email;
+
     @Pattern(regexp = "\\A(?!\\s*\\Z).+",
             message = "Login should be empty or with spaces")
+
     private String login;
     private String name;
     private String birthday;
